requirements.txt
python-dotenv
pandas 
google-api-python-client
s3fs